# PERF: avoid unnecessary reordering in MultiIndex._reorder_indexer